+++ Plutarch/Internal/TryFrom.hs
@@ -14,6 +14,7 @@ module Plutarch.Internal.TryFrom (
 ) where
 
 import Data.Functor.Const (Const)
+import Data.Kind (Type)
 import GHC.Generics (Generic)
 import Plutarch.Builtin.Bool (PBool, pif, (#||))
 import Plutarch.Builtin.ByteString (PByteString)
@@ -80,7 +81,7 @@ class PSubtype a b => PTryFrom (a :: PType) (b :: PType) where
 ptryFrom :: forall b a s r. PTryFrom a b => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r
 ptryFrom = ptryFrom'
 
-newtype Flip f a b = Flip (f b a) deriving stock (Generic)
+newtype Flip (f :: k1 -> k2 -> Type) (a :: k2) (b :: k1) = Flip (f b a) deriving stock (Generic)
 
 instance PTryFrom PData (PAsData PInteger) where
   type PTryFromExcess PData (PAsData PInteger) = Flip Term PInteger